Small edit to CPU profiling in main.c. On first pass the peak CPU loading and average CPU is reset to zero. Prevent 100% peak CPU loading being reported when first starting up.
Remove spi.o from Makefile. spi.c and spi.h not used.

Comment out program size text output in Makefile. Some of the tools / commands called may not be available or setup on Windows based machines. Commented out to avoid error being reported when compiling. Linux users with the appropriate tools setup can uncomment these lines to provide more information or feedback when compiling.

+++ src/planner.c
@@ -393,8 +393,8 @@ void planner_recalculate_trapezoids() {
       // Recalculate if current block entry or exit junction speed has changed.
       if (current->recalculate_flag || next->recalculate_flag) {
         // NOTE: Entry and exit factors always > 0 by all previous logic operations.
-        calculate_trapezoid_for_block(current, current->entry_speed/current->nominal_speed,
-          next->entry_speed/current->nominal_speed);
+        calculate_trapezoid_for_block(current, current->entry_speed/(float)(current->nominal_speed),
+          next->entry_speed/(float)(current->nominal_speed));
         current->recalculate_flag = 0; // Reset current only to ensure next trapezoid is computed
       }
     }
@@ -402,8 +402,8 @@ void planner_recalculate_trapezoids() {
   }
   // Last/newest block in buffer. Exit speed is set with MINIMUM_PLANNER_SPEED. Always recalculated.
   if(next != NULL) {
-    calculate_trapezoid_for_block(next, next->entry_speed/next->nominal_speed,
-      MINIMUM_PLANNER_SPEED/next->nominal_speed);
+    calculate_trapezoid_for_block(next, next->entry_speed/(float)(next->nominal_speed),
+      MINIMUM_PLANNER_SPEED/(float)(next->nominal_speed));
     next->recalculate_flag = 0;
   }
 }
